Extract tag correction 🐵patch from config.rb
